relay: Fix calculation of correction field
When calculating the correction field for follow-up frames we add the
peer delay to the residence time. The issue was that the peer delay was
added twice, once when we calculated the residence time and once when
setting the correction field.
Fix this by using only once the peer delay when calculating the
residence time.